requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
psutil>=5.9.0
python-dotenv>=1.0.0
